requirements-dev.txt
black
ruff
pre-commit
python-dotenv
